:recycle: Split commit questions to separate file
