+++ src/commands/commit/action.ts
@@ -1,31 +1,10 @@
-import inquirer, { QuestionCollection } from 'inquirer';
+import inquirer from 'inquirer';
 import { commitMessage } from './message';
+import { commitQuestions } from './questions';
 import { is } from 'typia';
-import { CommitMessageInput } from 'types/commit';
-
-type CommitMessageAnswers = Record<CommitMessageInput, string | undefined>;
-
-const questions: QuestionCollection<CommitMessageAnswers> = [
-  {
-    name: CommitMessageInput.TYPE,
-    type: 'input',
-  },
-  {
-    name: CommitMessageInput.SCOPE,
-    type: 'input',
-  },
-  {
-    name: CommitMessageInput.TITLE,
-    type: 'input',
-  },
-  {
-    name: CommitMessageInput.BODY,
-    type: 'input',
-  },
-];
 
 export const commitAction = async () => {
-  const answers: CommitMessageAnswers = await inquirer.prompt(questions);
+  const answers: CommitMessageAnswers = await inquirer.prompt(commitQuestions);
 
   Object.entries(answers).forEach(([key, value]) => {
     if (value && is<CommitMessageInput>(key)) {
